requirements.txt
absl-py==1.0.0
astunparse==1.6.3
bert-embedding==1.0.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Cython==0.29.23
filelock==3.4.0
flatbuffers==2.0
gast==0.4.0
gensim==4.1.2
gitdb==4.0.9
GitPython==3.1.24
gluonnlp==0.6.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.42.0
h5py==3.6.0
huggingface-hub==0.1.2
idna==2.6
importlib-metadata==4.8.2
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
lizard==1.17.9
Markdown==3.3.6
mxnet==1.4.0
nltk==3.6.5
numpy==1.17.0
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDriller==2.0
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
transformers==4.12.5
typing==3.6.6
typing_extensions==4.0.0
urllib3==1.22
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.13.3
zipp==3.6.0
